Refactor esgpull utils
